# backend/openapi.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/courses": {
            "get": {
                "operationId": "CoursesController_getCourses",
                "parameters": [
                    {
                        "name": "is_recommended",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "is_featured",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "category",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/courses/{courseId}/details": {
            "get": {
                "operationId": "CoursesController_getCourseDetails",
                "parameters": [
                    {
                        "name": "courseId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/courses/{courseId}/quizzes": {
            "get": {
                "operationId": "CoursesController_getCourseQuizzes",
                "parameters": [
                    {
                        "name": "courseId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/courses/purchasedCourses": {
            "post": {
                "operationId": "CoursesController_addPurchasedCourse",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUserPurchase"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            }
        },
        "/sections/{id}": {
            "get": {
                "operationId": "SectionsController_getSection",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/sections/{id}/test": {
            "get": {
                "operationId": "SectionsController_getTestBasedOnSectionId",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
...
        "/users/{userId}/courses": {
            "get": {
                "operationId": "UsersController_getUserCourses",
                "parameters": [
                    {
                        "name": "userId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
...
    },
    "info": {
        "title": "My API",
        "description": "API documentation",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "schemas": {
            "CreateUserPurchase": {
                "type": "object",
                "properties": {}
            },
...
        }
    }
}